+++ app/src/main/java/com/github/sdpsharelook/storage/WordSaver.kt
@@ -1,4 +1,6 @@
-package com.github.sdpsharelook
+package com.github.sdpsharelook.storage
+
+import com.github.sdpsharelook.Word
 
 class WordSaver(private val database: Database) : HashMap<String, Word>() {
     init {
